Refactor components, imports, SVGs, and folder structure
- Renamed MemberDate to MemberJoinDate and updated its import path.
- Moved SVG for MemberJoinDate to memberJoinSVG.
- Changed productList to productsList.
- Updated infoMessage component to ProductStatusMessage and renamed the file.
- Refactored TotalSales, moving its SVG to 	otalSellSvg.
- Updated import path for SelectProductQuantity to @/common/selectProductQuantity.
- Reorganized folder structure:
  - MemberJoinDate and TotalSales into subfolders with separate SVG files.
  - Updated paths for ProductStatusMessage and productsList.
- Applied minor style adjustments.

Ensure all components and styles are correctly integrated with the new structure.
