+++ app/admin/dashboard/components/unSoldProducts/index.js
@@ -1,4 +1,4 @@
-import ProductList from "../productList";
+import ProductsList from "../productsList";
 
 export const UnsoldProducts = ({ unsoldItems }) => {
   return (
@@ -6,13 +6,13 @@ export const UnsoldProducts = ({ unsoldItems }) => {
       {unsoldItems.length === 0 ? (
         ""
       ) : (
-        <div className="list">
+        <div className="list ">
           <h3 className="md:text-2xl font-heading py-2 text-xl font-bold ">
             Unsold Items ({unsoldItems.length})
           </h3>
           <div className="">
             {unsoldItems.map((item) => (
-              <ProductList key={item.id} unsoldData={item} />
+              <ProductsList key={item.id} unsoldData={item} />
             ))}
           </div>
         </div>
