+++ app/admin/dashboard/layout.js
@@ -4,7 +4,7 @@ import SideBar from "./components/sidebar";
 
 export default function DashboardLayout({ children }) {
   return (
-    <section className="flex md:gap-32  gap-4 md:pl-8 pl-4">
+    <section className="flex md:gap-20  gap-4 ">
       <SideBar />
       {children}
     </section>
