--- app/admin/dashboard/pendingorders/page.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-const PendingOrder = () => {
-  return <div>PendingOrder</div>;
-};
-
-export default PendingOrder;
