+++ app/product/[handle]/page.js
@@ -8,7 +8,7 @@ import BtnAddToCart from "@/common/btnAddToCart";
 import { Swiper, SwiperSlide } from "swiper/react";
 import "swiper/swiper-bundle.css";
 import { Navigation } from "swiper/modules";
-import SelectProductQuantity from "@/app/admin/dashboard/components/selectProductQuantity";
+import SelectProductQuantity from "@/common/selectProductQuantity";
 import { useParams } from "next/navigation";
 import AddProductReviews from "@/app/user/product/addProductReviews";
